Bug 622903 - fix coverity Defect Type: Code maintainability issues
https://bugzilla.redhat.com/show_bug.cgi?id=622903

Comment:
The "ptr" assignment on line 2622 is unnecessary because the value
is never used.

The "word" assignment on line 2630 can be removed because the first
token is ignored. Alternatively the value should be validated.

The "ptr" assignment on line 1456 is unnecessary because the value
is not used.

The "word" assignment on line 1463 can be removed because the first
token is ignored. Alternatively the value should be validated.

The "ptr" assignment on line 1210 is unnecessary because the value
is never used.

The "word" assignment on line 1217 can be removed because the first
token is ignored. Alternatively the value should be validated.