Bug 630093 - (cov#15518) Need to intialize fd in ldbm2ldif code
Currently, the ldbm_back_ldbm2ldif() function could bail due to an
error before fd is set.  We then attempt to close the file that fd
refers to.  We should initialize fd to STDOUT_FILENUM, as we skip
calling close() if fd is set to STDOUT_FILENUM.

Additionally, I noticed that we could call close() when fd is
negative or if it is STDERR or STDIN.  I fixed this so close() is
not called in those cases.