Bug 630096 - (cov#11778) check return value of ldap_parse_result
We were not checking the return value of ldap_parse_result in the
windows_check_user_password() function.  The old code was a bit
unclear about setting rc when we encountered errors from
ldap_result().  It also was calling ldap_parse_result() even if
ldap_result() encountered an error.  I fixed this code to be a
bit more straightforward.