Bug 630096 - (cov#15448) Check return value of cache_replace()
We need to check the return value of cache_replace() in
id2entry_add_ext().  The only possible error that can be returned
is when the entry we are trying to replace is not found in the
cache.  This should not occur since we are told that the entry
already exists by CACHE_ADD() just prior to this call.  If we run
into this situation, we will just log an error without adding the
entry to the cache.  This shouldn't be a big deal since the entry
will get added to the cache next time it is accessed.