Bug 630096 - (cov#15447) - Check return value of idl_append_extend()
We should check the return type of idl_append_extend(), though it does
not seem possible that the return type will be anything other than 0.
The only time idl_append_extend() returns anything other than 0 is
when it is unable to allocate memory.  Since the underlying allocation
function is slapi_ch_calloc(), the server will just exit if it runs
out of memory, which means we will never return up through
idl_append_extend().  The right thing to do from a code standpoint is
to still check for the return value though.