Bug 630094 - (cov#11818) Fix unreachable return in snmp subagent
The return statement at the end of agt_mopen_stats() is unreachable
according to coverity.  This return was removed before to fix the
coverity defect, but it was added back to fix a compiler warning.

We can satisfy both the compiler and coverity by adding a rc
variable to hold the return code. We can then return rc at the end
of the function. This also allows us to clean up all of the return
calls in this function by having all of them set rc and jump to a
label at the function end.