Bug 630094 - (cov#15452) Remove NULL checking for op_string
There is no chance for op_string to be NULL if "rc == LDAP_SUCCESS"
since op_string is set for all operation types in the switch statement,
which is the only place that rc can be set to something other that the
value of -1 that it is initialized to.

We can just skip the NULL checking for op_string in the calls to
slapi_log_error().  I also fixed the indentation to help me read
the code easier.