Bug 630094 - (cov#15455) Remove deadcode in attr_index_config()
If the index types (argv[1]) are not specified, attr_index_config()
bails.  We can remove some dead code where we check if "argc == 1"
later in the function since that case can never happen.

Additionally, we need to check if argc is 0, or if argv is NULL
before attempting to parse the list of attributes to be indexed.