Bug 630094 - (cov#15457) Remove deadcode in import code
The skipit variable is set to zero shortly before we check if it
is 0 in an if condition.  This if block can be removed since it
will never be hit.  The entry that was being freed in the if block
is already removed earlier in the function if skipit was non-0
prior to resetting skipit to 0.