Bug 630094 - (cov#15458) Fix deadcode issue in moddn code
In the moddn code that renames child entries, the for loop used to
rename the children can never be executed. Part of the condition
is that retval is 0, but retval will always be -1 the first time we
hit this loop. This only happens with subtree rename off, but it
should still be fixed.

The fix is to set retval to 0 at the prior to checking the condition
the first time.