Bug 630094 - (cov#15520) Fix unreachable code issue if perfctrs code
The final frees of priv->memory and priv will never be reached since
the function returns prior to these calls.  It looks as if an
"error:" label was removed at some point, as the WIN32 code in this
function has goto statements using that label, but the label is not
defined.

The fix is to add the "error:" label in ifdef blocks for  WIN32 that
calls the free of priv.  The free of priv->memory is not necessary
since WIN32 doesn't use it and non-WIN32 builds don't use the error
label at all.