Bug 630090 - (cov#15445) Fix illegal free in archive code
The directory variable points to a dynamically allocated string
returned by rel2abspath(). We are changing directory to point to
a string constant if we are unable to parse the directory. This
not only leaks memory, but it can cause us to attempt to free the
string constant.

We should free the string before we overwrite it, and we should
dynamically allocate a new string instead of using a string constant.