Bug 630091 - (cov#15514) Initialize DBT in entryrdn_get_parent()
In entryrdn_get_parent(), there is a DBT structure that we can use
without initializing.  If we goto the bail label, we try to free
data.data, but data was never initialized.  We should clear the
memory used by data in the beginning of the function before we have
an opportunity to goto bail.