Bug 630097 - (cov#11964) Remove dead code from libaccess
The libaccess library has some dead functions it it.  One of these
functions was flagged as having a NULL pointer dereference issue
by Coverity.  The problem function is unused, so it should be removed.
There are also a number of other unused functions in the same source
file that should be removed.