Bug 630097 - (cov#15461) Remove unnecessary NULL check in DNA
It is not necessary to check if config_entry->types is NULL since
it is guaranteed to be non-NULL by dna_parse_config_entry() when
it creates config_entry.  Coverity thinks that a NULL derefence is
possible since we are checking if config_entry->types is NULL.  We
should remove this NULL check.