Bug 630097 - (cov#15463) Remove NULL check in referint plugin
Coverity believes that search_result_pb can be NULL since we check
if it is NULL before freeing the internal search results.  If this
was true, there would be a NULL dereference issue when we call
slapi_pblock_get().  We are guaranteed that search_result_pb is
non-NULL after slapi_pblock_new() is called since the server would
exit if it was unable to allocate memory.

We should remove the NULL check before freeing the internal search
results.