Bug 630097 - (cov#15464) NULL dereference in repl code
If the attr parameter that is passed to my_ber_scanf_attr() is
NULL, we jump to the loser label where we clean up memory we may
have allocated.  We dereference attr without first checking if it
is NULL in this clean-up code.  We need to check if attr is NULL
before dereferencing it.