Bug 630097 - (cov#15507,15508) NULL dereference in entryrdn code
In entryrdn_compare_dups(), we dereference the a and b parameters
when initializing the elem_a and elem_b variables.  We later
perform NULL checks on both a and b, but a NULL would have
triggered a crash.  We should not dereference a or b until after
the NULL checks are performed.