Bug 630091 - (cov#11973) Array overrun in libaccess
When going through the exceptions table in libaccess, we don't
check if we are at the last pair of elements in the array before
incrementing to the next pair.  This patch adds checks to see if
we are at the last pair of elements and avoids the increment if
necessary.