Bug 522055 - Scope check for managed attribute fails
When using the linked attribute plug-in, an entry that is renamed
that is outside of the scope of the plug-in will cause the forward
links in other entries to be updated if they contain a managed
attribute type.

We need to check if the new DN of the renamed entry is within the
scope of the configured linked attributes before updating forward
links.