Bug 625335 - Self-write aci has permission to invalid attribute
The Console throws an error dialog when you attempt to edit the
default self-write ACI.  The Console thinks that the labeledURL
attribute is not valid sicne it does not properly handle attributes
that use anything other than the primary short name.

This patch changes the default ACI to use the primary short name
of labeledURI instead of labeledURL.  Fixing the Console to handle
non-primary short names will be dealt with in another bug in the
future.