Bug 640854 - changelog db: _cl5WriteOperation: failed to write entry; db error - 22 Invalid argument
https://bugzilla.redhat.com/show_bug.cgi?id=640854

Description: DBENV open flags is used to determine the DB_OPEN mode
whether to set DB_AUTO_COMMIT or not.  The info was eliminated in
the change made for "Bug 633168 - Share backend dbEnv with the
replication changelog".

This patch picks up the backend dbenv openflags and uses it for
the changelog DB_OPEN.