Bug 544321 - remove-ds.pl should not throw error unlabelling port
When removing an instance using remove-ds.pl, a fatal error will
be thrown when trying to remove the SELinux port label if the port
is not labelled.  This patch makes this case a non-error since
there is no need to complain about removing a label if it has
already been removed.