Bug 555955 - Allow CoS values to be merged
This patch adds the ability for CoS values to be merged and create
multi-valued attributes.  One can append "merge-schemes" the the
end of the cosAttribute value in a definition entry to allow values
to be merged.

With a single indirect CoS definition, a merge will make CoS use
each specifier attribute value in the target entry to look for
CoS values in each of the found template entries.  All of these
values will then be applied to the target entry (with the exception
of duplicate values).

With multiple indirect CoS definitions for the same attribute,
setting merge mode for all definitions will cause all of the
definitions to be used to find the values to apply to the target
entry.  If merge-schemes is not defined for all of these definitions,
the result is undefined (values from the first found CoS definition
will be applied).