Bug 619633 - Make attribute uniqueness obey requiredObjectClass
The attribute uniqueness plug-in currently does not enforce the
requiredObjectClass setting properly.  The plug-in checks that the
entry being modified meets the requiredObjectClass setting, but it
does not check if the potential conflicts in the database meet this
requirement as well.

This patch changes the behavior to check that both the target and
potential conflicting entries meet the requiredObjectClass setting.