Bug 644608 - RHDS 8.1->8.2 upgrade fails to properly migrate ACIs
https://bugzilla.redhat.com/show_bug.cgi?id=644608

Description: Adding the ancestorid fix code to ##upgradednformat.pl.
The upgrade utility upgradednformat in the previous version could
have broken the ancestorid index file.  The ID list in the updated
index value may not be in the ascendant order.  This patch checks
the ID order in the ID lists.  Once it detects the disorder, it
reindexes the corrupted ancestorid.