Bug 619623 - attr-unique-plugin ignores requiredObjectClass on modrdn operations
The attribute uniqueness plug-in is not checking if renamed entries
meet the requiredObjectClass requirements.  In addition, a rename
that uses a new superior is not properly handled.

This patch adds a check for requiredObjectClass for MODRDN ops.  I
also added a helper function called slapi_entry_rename() which allows
the caller to apply a MODRDN operation to a Slapi_Entry.  My patch
makes use of this to create a renamed dummy entry that is used to
check for any attribute uniqueness conflicts.