Bug 638773 - permissions too loose on pid and lock files
https://bugzilla.redhat.com/show_bug.cgi?id=638773

Description: This patch changes the permissions of the pid and
lock files to be writable only by the owner (0644).  They should
remain readable by all since they may serve in an advisory role
to other processes that need to determine if it's running.

Checking in this patch on behalf of Ulf Weltman (ulf.weltman@hp.com).