Bug 647932 - multiple memberOf configuration adding memberOf where there is no member
There is a bug in the way we construct the filter in memberof_call_foreach_dn()
when multiple grouping attribtues are set.  We should be constructing a filter
using the passed in types and dn that looks like this:

  (|(type1=dn)(type2=dn))

Instead, we have hard-coded wildcards in the place of the dn when multiple
types are passed in.  The result is a filter that looks like this:

  (|(type1=*)(type2=*))

When this function is used to find groups who have dn as a member, we end up
finding all groups that have any grouping attribute present and treat dn as if
it is a member.

This issue does not occur when a single type is used.