Bug 521088 - DNA should check ACLs before getting a value from the range
The DNA plug-in gets a value from the range at the pre-op stage, which
means that the operation can fail at a later point.  This results in
a value being used up from the range even though it is never actually
used within an entry.

This patch makes DNA check the ACIs to see if the incoming operation
is allowed before allocating a value from the range.  If the ACIs
result in an operation that will be rejected, we skip allocation
which keeps us from wasting a value from the range.