Bug 567282 - server can not abandon searchRequest of "simple paged results"
https://bugzilla.redhat.com/show_bug.cgi?id=567282

Description: Simple Paged Results search keeps the connection
per paging, but not an operation.  When an abandon request is
issued, the operation referred by the request has already finished.
This patch introduces pagedresults_cleanup function to check
whether the connection is for the simple paged results or not, and
if it is, the simple paged results is cleaned up.  If it is not,
pagedresults_cleanup does nothing.  The function is called from
do_abandon as well as from connection_cleanup.