Bug 656392 - Remove calls to ber_err_print()
The LDAPDebug macros call the ber_err_print() function in cases
where SLAPD_LOGGING is not defined.  This function exists as a
private function in MozLDAP, but it does not exist in the OpenLDAP
client libraries.  This results in the server aborting when built
against OpenLDAP due to an unresolved symbol.

This patch removes the code that calls ber_err_print().  I removed
the entire SLAPD_LOGGING define from the code, as it was only used
to switch between logging to stderr or the errors log subsystem.
This code was very old.  We should only be logging to the errors
log subsystem.