Bug 656515 - Allow Name and Optional UID syntax for grouping attributes
The config validation code in the memberOf plug-in currently requires that the
grouping attribute be defined to use the Distinguished Name syntax.  The
uniqueMember attribute is still a common method of grouping users, but this
attribute is defined to use the Name and Optional UID syntax.  This syntax
contains a DN plus an optional hex-valued UID that can be appended to the end.

We should allow attribute defined to use the Name and Optional UID syntax to be
used as memberOf grouping attributes.  We will not support one actually using
the optional UID portion in the value, but this is extremely rare in practice.