Bug 197886 - Avoid overflow of UUID generator
It is theoretically possible to overflow the UUID generator if you
generate more than 10M UUIDs in a single minute.  Though this is
a highly unlikely scenario, we should avoid creating any duplicate
UUIDs.  This patch makes the server reject any ADD operations or
import tasks that would overflow the UUID generator.