Bug 197886 - Initialize return value for UUID generation code
The previous patch for bug 197886 had a bug in it where the return
value of import_generate_uniqueid() is not initialized.  If this
function is called when a UUID is not needed, the return value is
garbage.  This causes the bulk import code on a replica that is
being initialized to abort if garbage is returned.