Bug 658312 - Invalid free in Managed Entry plug-in
The previous patch for bug 658312 would perform an invalid free
with a mapping rule used the $dn macro.  We get the DN of the
origin entry with slapi_entry_get_ndn().  The returned value is
not supposed to be freed.  This patch adds a flag to indicate it
we need to perform a free of the mapping value or not.