Bug 641944 - Don't normalize non-DN RDN values
Our current DN normalization code assumes that a RDN value can
contain a DN pointer.  It also normalizes this RDN value as a DN,
regardless of the syntax of the RDN attribute type.  This causes
non-DN RDN values to get messed up.

This patch makes the DN normalization code only normalize a RDN
value as a DN for an attribute type that is defined to use the
Distinguished Name syntax.