Bug 606439 - Creating server instance with LDAPI takes too long
The setup-ds.pl script expects ns-slapd to write a "slapd started"
message to the errors log when the server comes up.  When you
create an instance that does not listen on a standard LDAP port,
the "slapd started" string is never printed.  This causes the
installer to hang for 10 minutes, at which point in prints an error
that it could not create the instance.

This patch ensures that the "slapd started" message is printed
for servers that only listen on LDAPS or LDAPI interfaces.