Bug 632670 - Chain-on-update logs managed-entries-plugin errors
When using chain-on-update, SLAPI_ENTRY_POST_OP will not be set on
the consumer for a post-modify callback.  The managed entry plug-in
currently logs a fatal message if it is unable to fetch the post-op
entry.

This patch simply changes the message to the SLAPI_LOG_PLUGIN
level.  This is fine as we have no work to do if we do not have a
post-op entry.