Bug 663191 - Don't use $USER in DSCreate.pm
The $USER environment variable is used in the instance creation
and removal code to determine what location to use for initconfig
files.  This will not work correctly if $USER is not set, or if
one used su prior to running the installer.

This patch uses DSUtil::getLogin instead, which gets the login
name without using the $USER environment variable.